requirements.txt
Django==1.5.1
Markdown==2.3.1
django-cors-headers==0.12
django-filter==0.6
dj-database-url==0.2.2
dj-static==0.0.5
djangorestframework==2.3.6
gunicorn==17.5
psycopg2==2.4.5